+++ docling/models/inference_engines/vlm/_utils.py
