+++ docling/models/inference_engines/vlm/api_openai_compatible_engine.py
@@ -9,11 +9,11 @@ from typing import TYPE_CHECKING, List, Optional
 from PIL.Image import Image
 
 from docling.datamodel.vlm_engine_options import ApiVlmEngineOptions
-from docling.models.runtimes._utils import (
+from docling.models.inference_engines.vlm._utils import (
     extract_generation_stoppers,
     preprocess_image_batch,
 )
-from docling.models.runtimes.base import (
+from docling.models.inference_engines.vlm.base import (
     BaseVlmEngine,
     VlmEngineInput,
     VlmEngineOutput,
