+++ docling/models/inference_engines/vlm/base.py
