--- docling/models/runtimes/vlm/__init__.py
@@ -1,15 +1,0 @@
-"""VLM model family inference engines."""
-
-from docling.models.runtimes.vlm.api_openai_compatible_engine import ApiVlmEngine
-from docling.models.runtimes.vlm.auto_inline_engine import AutoInlineVlmEngine
-from docling.models.runtimes.vlm.mlx_engine import MlxVlmEngine
-from docling.models.runtimes.vlm.transformers_engine import TransformersVlmEngine
-from docling.models.runtimes.vlm.vllm_engine import VllmVlmEngine
-
-__all__ = [
-    "ApiVlmEngine",
-    "AutoInlineVlmEngine",
-    "MlxVlmEngine",
-    "TransformersVlmEngine",
-    "VllmVlmEngine",
-]
