+++ tests/test_vlm_presets_and_runtime_options.py
@@ -30,7 +30,7 @@ from docling.datamodel.vlm_engine_options import (
     TransformersVlmEngineOptions,
     VllmVlmEngineOptions,
 )
-from docling.models.runtimes.base import VlmEngineType
+from docling.models.inference_engines.vlm import VlmEngineType
 
 # =============================================================================
 # RUNTIME OPTIONS TESTS
